Reduce latency of collapsible
From my understanding, if you collapse all items at the beginning, then you don't need the "time out" part. Therefore it will reduce that 10ms delay when user expand or collapse items. It seems to be effective and working when I tested it on my local host. Try it yourself!